requirements.txt
aiogram==3.13.1
httpx==0.27.2
apscheduler==3.10.4
python-dotenv==1.0.1
